requirements.txt
Django==1.7.1
Markdown==2.5.1
django-auth-ldap==1.2.2
django-filter==0.8
djangorestframework==2.4.4
python-ldap==2.4.18
requests==2.4.3
